Finished form splitting and splitting endpoints
