+++ myproject/app/cookie/__init__.py
@@ -1,0 +1,5 @@
+from flask import Blueprint
+
+cookie_bp = Blueprint("cookie", __name__, template_folder="templates")
+
+from . import controller
