+++ myproject/app/feedback/__init__.py
@@ -1,0 +1,5 @@
+from flask import Blueprint
+
+feedback_bp = Blueprint("feedback", __name__, template_folder="templates")
+
+from . import controller
