+++ myproject/app/feedback/templates/feedback.html
@@ -22,7 +22,7 @@
     </style>
     <div class="largediv">
     <h1>Publish feedback</h1>
-    <form action="{{ url_for('add_feedback') }}" method="post" class="w-50 mx-auto gap-4">
+    <form action="{{ url_for('feedback.add_feedback') }}" method="post" class="w-50 mx-auto gap-4">
         {{ form.csrf_token() }}
         <div class="mb-3">
             {{ render_field(form.feedback, class='form-control') }}
