+++ myproject/app/general/templates/skills.html
