+++ myproject/app/todo/templates/add_task.html
@@ -19,7 +19,7 @@
     <div class="largediv">
         
         <h1>Add Task</h1>
-        <form action="{{ url_for('add_task') }}" method="post"
+        <form action="{{ url_for('todo.add_task') }}" method="post"
               class="row d-flex justify-content-center w-75 gap-1">
             {{ form.csrf_token() }}
             <div class="row d-flex justify-content-center">
