+++ myproject/app/todo/templates/todo.html
@@ -23,7 +23,7 @@
     <h1>Todo</h1>
     <div class="container-lg w-70">
         {% if session.get('user') != None %}
-    <a href="{{ url_for('add_task') }}" role="button" tabindex="-1" class="w-30 mx-auto mt-2 btn btn-secondary">Add Task</a>
+    <a href="{{ url_for('todo.add_task') }}" role="button" tabindex="-1" class="w-30 mx-auto mt-2 btn btn-secondary">Add Task</a>
 {% endif %}
         <table class="table">
             <thead>
@@ -44,14 +44,14 @@
                     <td>{{ task.status.value }}</td>
                     {% if session.get('user') != None %}
                         <td>
-                            <a href="{{ url_for('update_task', id=task.id) }}">
+                            <a href="{{ url_for('todo.update_task', id=task.id) }}">
                                 <button class="btn btn-blue">
                                     Update
                                 </button>
                             </a>
                         </td>
                         <td>
-                            <a href="{{ url_for('delete_task', id=task.id) }}">
+                            <a href="{{ url_for('todo.delete_task', id=task.id) }}">
                                 <button class="btn btn-blue">
                                     Delete
                                 </button>
